Created trait and change entity, listener
