+++ app/config/config.yml
@@ -59,13 +59,7 @@ doctrine:
         auto_generate_proxy_classes: "%kernel.debug%"
         naming_strategy: doctrine.orm.naming_strategy.underscore
         auto_mapping: true
-        mappings:
-          translatable:
-              type: annotation
-              alias: Gedmo
-              prefix: GedmoTranslatableEntity
-              # make sure vendor library location is correct
-              dir: "%kernel.root_dir%/../vendor/gedmo/doctrine-extensions/lib/Gedmo/Translatable/Entity"
+
 
 # Swiftmailer Configuration
 swiftmailer:
