+++ src/AppBundle/Entity/Framework.php
@@ -6,6 +6,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Validator\Constraints as Assert;
+use AppBundle\Traits\AuthoreditorEntity;
 
 /**
  * Framework
@@ -48,6 +49,8 @@ class Framework
      * @ORM\OneToMany(targetEntity="Site", mappedBy="framework")
      */
     private $sites;
+
+    use AuthoreditorEntity;
 
     public function __construct()
     {
