+++ src/AppBundle/Listener/SiteAuthor.php
@@ -3,25 +3,24 @@
 namespace AppBundle\Listener;
 
 use Doctrine\ORM\Event\LifecycleEventArgs;
-use Symfony\Component\Security\Core\SecurityContextInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 class SiteAuthor
 {
 
-    private $security;
+    protected $container;
 
-    public function __construct(SecurityContextInterface $security) {
-        $this->security = $security;
+    public function __construct(ContainerInterface $container = null) {
+        $this->container = $container;
     }
 
-    public function postUpdate(LifecycleEventArgs $args) {
+    public function preUpdate(LifecycleEventArgs $args) {
         $entity = $args->getEntity();
         $entityManager = $args->getEntityManager();
-        $user = $this->security->getToken()->getUser();
-        if ($entity instanceof \AppBundle\Entity\Site) {
+        $user = $this->container->get('security.context')->getToken()->getUser();
+        if (!$entity instanceof \AppBundle\Entity\User) {
             $user = $entityManager->getRepository('AppBundle\Entity\User')->find($user->getId());
             $entity->setEditor($user);
-            $entityManager->flush();
         }
     }
 
@@ -29,12 +28,11 @@ class SiteAuthor
     public function prePersist(LifecycleEventArgs $args) {
         $entity = $args->getEntity();
         $entityManager = $args->getEntityManager();
-        $user = $this->security->getToken()->getUser();
-        if ($entity instanceof \AppBundle\Entity\Site) {
+        $user = $this->container->get('security.context')->getToken()->getUser();
+        if (!$entity instanceof \AppBundle\Entity\User) {
             $user = $entityManager->getRepository('AppBundle\Entity\User')->find($user->getId());
             $entity->setAuthor($user);
             $entity->setEditor($user);
-            $entityManager->flush();
         }
     }
